# Optimize SpecialBracketCharacters parsing and initialization of master caches