--- a/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/Xaml/Parser/SpecialBracketCharacters.cs
+++ b/src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/Xaml/Parser/SpecialBracketCharacters.cs
@@ -7,10 +7,11 @@
 //  Description: Data model for the Bracket characters specified on a Markup Extension property.
 
 using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.Text;
+using System.Buffers;
+using System.Windows.Markup;
+using System.ComponentModel;
+using System.Collections.Generic;
 
 namespace MS.Internal.Xaml.Parser
 {
@@ -20,6 +21,118 @@ namespace MS.Internal.Xaml.Parser
     /// </summary>
     internal sealed class SpecialBracketCharacters : ISupportInitialize
     {
+#if !NETFX
+        /// <summary>
+        /// Stores characters that cannot be specified in <see cref="MarkupExtensionBracketCharactersAttribute"/>.
+        /// </summary>
+        private static readonly SearchValues<char> s_restrictedCharSet = SearchValues.Create('=', ',', '\'', '"', '{', '}', '\\');
+
+        private string _startChars;
+        private string _endChars;
+        private bool _initializing;
+        private StringBuilder _startCharactersStringBuilder;
+        private StringBuilder _endCharactersStringBuilder;
+
+        internal SpecialBracketCharacters()
+        {
+            BeginInit();
+        }
+
+        internal SpecialBracketCharacters(IReadOnlyDictionary<char, char> attributeList)
+        {
+            BeginInit();
+            if (attributeList != null && attributeList.Count > 0)
+            {
+                Tokenize(attributeList);
+            }
+        }
+
+        internal void AddBracketCharacters(char openingBracket, char closingBracket)
+        {
+            if (_initializing)
+            {
+                _startCharactersStringBuilder.Append(openingBracket);
+                _endCharactersStringBuilder.Append(closingBracket);
+            }
+            else
+            {
+                throw new InvalidOperationException();
+            }
+        }
+
+        private void Tokenize(IReadOnlyDictionary<char, char> attributeList)
+        {
+            if (_initializing)
+            {
+                foreach (char openingBracket in attributeList.Keys)
+                {
+                    char closingBracket = attributeList[openingBracket];
+
+                    if (IsValidBracketCharacter(openingBracket, closingBracket))
+                    {
+                        _startCharactersStringBuilder.Append(openingBracket);
+                        _endCharactersStringBuilder.Append(closingBracket);
+                    }
+                }
+            }
+        }
+
+        private static bool IsValidBracketCharacter(char openingBracket, char closingBracket)
+        {
+            if (openingBracket == closingBracket)
+                throw new InvalidOperationException("Opening bracket character cannot be the same as closing bracket character.");
+            else if (char.IsLetterOrDigit(openingBracket) || char.IsLetterOrDigit(closingBracket) || char.IsWhiteSpace(openingBracket) || char.IsWhiteSpace(closingBracket))
+                throw new InvalidOperationException("Bracket characters cannot be alpha-numeric or whitespace.");
+            else if (s_restrictedCharSet.Contains(openingBracket) || s_restrictedCharSet.Contains(closingBracket))
+                throw new InvalidOperationException("Bracket characters cannot be one of the following: '=' , ',', '\'', '\"', '{ ', ' }', '\\'");
+
+            return true;
+        }
+
+        internal bool IsSpecialCharacter(char ch)
+        {
+            return _startChars.Contains(ch) || _endChars.Contains(ch);
+        }
+
+        internal bool StartsEscapeSequence(char ch)
+        {
+            return _startChars.Contains(ch);
+        }
+
+        internal bool EndsEscapeSequence(char ch)
+        {
+            return _endChars.Contains(ch);
+        }
+
+        internal bool Match(char start, char end)
+        {
+            return _endChars.IndexOf(end, StringComparison.Ordinal) == _startChars.IndexOf(start, StringComparison.Ordinal);
+        }
+
+        internal string StartBracketCharacters
+        {
+            get { return _startChars; }
+        }
+
+        internal string EndBracketCharacters
+        {
+            get { return _endChars; }
+        }
+
+        public void BeginInit()
+        {
+            _initializing = true;
+            _startCharactersStringBuilder = new StringBuilder();
+            _endCharactersStringBuilder = new StringBuilder();
+        }
+
+        public void EndInit()
+        {
+            _startChars = _startCharactersStringBuilder.ToString();
+            _endChars = _endCharactersStringBuilder.ToString();
+            _initializing = false;
+        }
+#else
         private string _startChars;
         private string _endChars;
         private readonly static ISet<char> _restrictedCharSet = new SortedSet<char>((new char[] { '=', ',', '\'', '"', '{', '}', '\\' }));
@@ -134,5 +247,6 @@ namespace MS.Internal.Xaml.Parser
             _endChars = _endCharactersStringBuilder.ToString();
             _initializing = false;
         }
+#endif
     }
 }